moved `importIdentifier` to `Importer`
