+++ utils/src/main/scala/org/alephium/ralph/lsp/utils/URIUtil.scala
@@ -152,25 +152,4 @@ object URIUtil {
     )
   }
 
-  /**
-   * String literal that defines an import statement for a source file.
-   *
-   * @return If the file is named `std/my_code.ral`, the import statement returned
-   *         is `import "std/my_code"`.
-   */
-  def importIdentifier(uri: URI): Option[String] =
-    takeRight(
-      uri = uri,
-      count = 2
-    ) map {
-      identifier =>
-        val string =
-          identifier
-            .iterator()
-            .asScala
-            .mkString("/") // import statements use forward slash.
-
-        string.substring(0, string.lastIndexOf("."))
-    }
-
 }
